+++ R/zzz_utils.R
@@ -1,14 +1,13 @@
 #NAMESPACE ----------
 
-#' @importFrom dplyr all_of any_of arrange as_tibble bind_cols bind_rows desc distinct everything filter group_by group_split mutate pull relocate rename rename_with select slice_max slice_min starts_with ungroup vars
-#' @importFrom ggplot2 %+replace% aes coord_cartesian element_rect facet_grid facet_wrap labeller geom_area geom_histogram geom_hline geom_line geom_point geom_rug geom_segment geom_vline ggplot label_both labs theme_bw scale_x_continuous scale_y_continuous scale_shape_manual scale_color_manual scale_linetype_manual stat_function theme
+#' @importFrom dplyr across all_of any_of arrange as_tibble bind_cols bind_rows case_when desc distinct everything filter full_join group_by group_split mutate pull relocate rename rename_with row_number select slice slice_max slice_min starts_with summarise ungroup vars
+#' @importFrom ggplot2 %+replace% aes coord_cartesian element_rect facet_grid facet_wrap labeller geom_area geom_histogram geom_hline geom_line geom_point geom_ribbon geom_rug geom_segment geom_vline ggplot label_both labs theme_bw scale_x_continuous scale_y_continuous scale_y_log10 scale_fill_manual scale_shape_manual scale_color_manual scale_linetype_manual stat_function theme
 #' @importFrom magrittr %>%
-#' @importFrom mrgsolve as.list data_set ev is.mrgmod mread mcode mrgsim mrgsim_df mrgsim_q mvgauss obsaug omat param realize_addl smat zero_re
+#' @importFrom mrgsolve as.list collapse_omega data_set ev is.mrgmod mread mcode mrgsim mrgsim_df mrgsim_q mvgauss omat param realize_addl smat zero_re
 #' @importFrom optimx optimx
-#' @importFrom purrr map map2 map_dbl map2_dbl map_dfr pmap pmap_dfr quietly simplify transpose
+#' @importFrom purrr map map2 map_dbl map_dfr map_lgl map2_dbl map2_dfc pmap pmap_dfr quietly safely simplify transpose
 #' @importFrom rlang .data set_names
-#' @importFrom stats runif dnorm pnorm qnorm
-#' @importFrom stringr str_subset str_detect str_which str_replace
+#' @importFrom stringr str_detect str_extract_all str_replace str_subset str_which
 #' @importFrom tibble tibble rownames_to_column
 #' @importFrom tidyr expand_grid fill pivot_longer
 NULL
@@ -36,7 +35,7 @@ odiag <- function(x){
 #' @noRd
 get_quantile <- function(x, .p){
   if(!is.mrgmod(x)) stop("the first argument to lowbounds must be a model object", call. = F)
-  map_dbl(sqrt(odiag(x)), qnorm, p = .p, mean = 0)
+  map_dbl(sqrt(odiag(x)), stats::qnorm, p = .p, mean = 0)
 }
 
 #' Internal "mapbayr" model examples
@@ -55,4 +54,14 @@ my_percent <- function(x){
 eta_from_opt <- function(x){
   stopifnot(is.data.frame(x))
   unlist(x[,grepl("ETA", names(x))])
+}
+
+ci2q <- function(ci) (1-(ci/100))/2
+znorm <- function(ci){
+  stopifnot(is.numeric(ci), ci > 0, ci < 100)
+  stats::qnorm(1-ci2q(ci))
+}
+
+namephicov <- function(n){
+  unlist(map(seq_len(n), ~ paste0("ETC",.x,"_",unlist(combn(.x, 1, simplify = FALSE)))))
 }